requirements.txt
altgraph==0.17.3
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blurhash==1.1.4
boost==0.1
Brotli @ file:///D:/bld/brotli-split_1693583621767/work
cachetools==4.2.4
certifi @ file:///C:/b/abs_91u83siphd/croot/certifi_1700501720658/work/certifi
cffi==1.16.0
channels==4.0.0
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1688813409104/work
cmake==3.22.4
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.0
cycler==0.11.0
DBUtils==3.1.0
decorator==4.4.2
defusedxml==0.7.1
Django==4.2
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-filter==24.1
django-js-asset==2.2.0
djangorestframework==3.14.0
docopt==0.6.2
easydict==1.10
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.1
flatbuffers==24.3.7
fonttools==4.42.1
gast==0.4.0
gevent==24.2.1
google-pasta==0.2.0
greenlet==3.0.3
humanfriendly==10.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib-resources==6.0.1
importlib_metadata==7.1.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
labelImg==1.8.6
lapx==0.5.4
MarkupSafe==2.1.5
Mastodon.py==1.8.0
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.2
moviepy==1.0.3
mpmath==1.3.0
mysqlclient==2.2.4
natsort==8.3.1
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nodejs==0.1.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.1
opencv-python==4.8.0.76
openpyxl==3.1.2
opt-einsum==3.3.0
optional-django==0.1.0
packaging==23.1
pandas==2.1.0
pandocfilters==1.5.1
parso==0.8.4
pefile==2023.2.7
pickleshare==0.7.5
Pillow==10.0.0
pipreqs==0.5.0
platformdirs==4.2.0
prettytable==3.10.0
proglog==0.1.10
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.5
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycparser==2.21
pyecharts==2.0.5
Pygments==2.17.2
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
PyMySQL==1.1.0
pyparsing==3.0.9
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyqt5-tools==5.15.9.3.3
pyreadline3==3.4.1
PySide6==6.5.2
PySide6-Addons==6.5.2
PySide6-Essentials==6.5.2
PySocks @ file:///D:/bld/pysocks_1661604991356/work
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.2
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
referencing==0.34.0
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
requests-oauthlib==1.3.1
rpds-py==0.18.0
rsa==4.9
schedule==1.2.1
scipy==1.11.2
seaborn==0.12.2
shiboken6==6.5.2
simplejson==3.19.2
six==1.15.0
snapshot-phantomjs==0.0.3
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
tinycss2==1.2.1
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1688315532570/work
tzdata==2023.3
ultralytics==8.0.169
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1689789803562/work
wcwidth==0.2.13
webencodings==0.5.1
websocket==0.2.1
websocket-client==0.57.0
win-inet-pton @ file:///D:/bld/win_inet_pton_1667051142467/work
wrapt==1.12.1
yarg==0.1.9
zipp==3.16.2
zope.event==5.0
zope.interface==6.2
